+3 tasks. 19.11.2021.

+++ src/main/java/strings/ReverseInteger1.java
@@ -3,14 +3,14 @@ package strings;
 public class ReverseInteger1 {
 
     public static int reverse(int x) {
-        int result = 0;
+        long result = 0;
 
         while (x != 0) {
             result = result * 10 + x % 10;
             if (result < Integer.MIN_VALUE || result > Integer.MAX_VALUE) return 0;
             x /= 10;
         }
-        return result;
+        return (int) result;
     }
 
     public static void main(String[] args) {
